:sparkles: Add function to modify the WP nav menu classes
We don't want the standard blog page to be identified in the nav
menu as the parent of the courses CPT posts. This function filters
the nav menu item classes array to remove the "active" class from
the blog home page when viewing any Courses CPT page or post, and
reassigns it to the 'post_type_archive' nav menu item.